requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
future==0.17.1
importlib-metadata==0.18
logzero==1.5.0
path.py==12.0.1
pycparser==2.19
python-telegram-bot==11.1.0
six==1.12.0
youtube-dl==2019.6.8
zipp==0.5.1
